rename backend and internal dirs
