+++ internal/delivery/http/handlers/course_handlers_test.go
@@ -5,10 +5,10 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"skillForce/backend/delivery/http/response"
-	"skillForce/backend/models"
-	"skillForce/backend/repository/mock"
-	"skillForce/backend/usecase"
+	"skillForce/internal/delivery/http/response"
+	"skillForce/internal/models"
+	"skillForce/internal/repository/mock"
+	"skillForce/internal/usecase"
 	"strings"
 	"testing"
 )
