+++ internal/models/course.go
